# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WorkForcePro API",
    "version": "1.0.0",
    "description": "RESTful API for managing employees, trainings, certifications, and departments in a workforce management system."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/employees/{id}/certifications": {
      "get": {
        "tags": ["Employees", "Certifications"],
        "summary": "Get certifications for an employee",
        "operationId": "getEmployeeCertifications",
        "parameters": [
          { "$ref": "#/components/parameters/EmployeeId" }
        ],
        "responses": {
          "200": {
            "description": "Employee certifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Certification" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/certifications": {
      "get": {
        "tags": ["Certifications"],
        "summary": "List all certifications",
        "operationId": "getCertifications",
        "responses": {
          "200": {
            "description": "List of certifications with employee names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Certification" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Certifications"],
        "summary": "Create a new certification for an employee",
        "operationId": "createCertification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCertificationRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Certification created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Certification" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/certifications/expiring": {
      "get": {
        "tags": ["Certifications"],
        "summary": "Get expiring certifications (within 30 days)",
        "operationId": "getExpiringCertifications",
        "responses": {
          "200": {
            "description": "List of expiring certifications (or all certifications if none expiring - known bug)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Certification" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/certifications/employee/{employee_id}": {
      "get": {
        "tags": ["Certifications"],
        "summary": "Get certifications for a specific employee",
        "operationId": "getCertificationsByEmployee",
        "parameters": [
          { "$ref": "#/components/parameters/EmployeeNumericId" }
        ],
        "responses": {
          "200": {
            "description": "Employee certifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Certification" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/certifications/{id}": {
      "get": {
        "tags": ["Certifications"],
        "summary": "Get a single certification",
        "operationId": "getCertification",
        "parameters": [
          { "$ref": "#/components/parameters/CertificationId" }
        ],
        "responses": {
          "200": {
            "description": "Certification details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Certification" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "put": {
        "tags": ["Certifications"],
        "summary": "Update a certification",
        "operationId": "updateCertification",
        "parameters": [
          { "$ref": "#/components/parameters/CertificationId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCertificationRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Certification updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Certification" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "tags": ["Certifications"],
        "summary": "Delete a certification",
        "operationId": "deleteCertification",
        "parameters": [
          { "$ref": "#/components/parameters/CertificationId" }
        ],
        "responses": {
          "204": { "description": "Certification deleted, no content" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "EmployeeId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "Employee database ID"
      },
      "EmployeeNumericId": {
        "name": "employee_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "Employee database ID"
      },
...
      "CertificationId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "Certification database ID"
      }
    },
    "schemas": {
...
      "Certification": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "employee_id": { "type": "integer", "example": 2 },
          "employee_name": { "type": "string", "example": "John Doe" },
          "certification_name": { "type": "string", "example": "AWS Solutions Architect" },
          "expiry_date": { "type": "string", "format": "date", "example": "2026-12-31" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateCertificationRequest": {
        "type": "object",
        "required": ["employee_id", "certification_name", "expiry_date"],
        "properties": {
          "employee_id": { "type": "integer", "example": 2 },
          "certification_name": { "type": "string", "example": "AWS Solutions Architect" },
          "expiry_date": { "type": "string", "format": "date", "example": "2026-12-31" }
        }
      },
      "UpdateCertificationRequest": {
        "type": "object",
        "required": ["certification_name", "expiry_date"],
        "properties": {
          "certification_name": { "type": "string", "example": "AWS Solutions Architect - Professional" },
          "expiry_date": { "type": "string", "format": "date", "example": "2027-06-30" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Employee not found" }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  }
}